modification done by tl

+++ GitProj1/src/main/java/com/nit/branch1/UPIPayment.java
@@ -2,7 +2,7 @@ package com.nit.branch1;
 
 public class UPIPayment {
 	
-	public String showBalance(String upiId,float amount)
+	public String showBalance(String upiId,int amount)
 	{
 		return "your balance: "+amount;
 	}
